Move Priority to constants.
